fix: Update functionality according to TODOs
