+++ core/libs/api/src/lib/model-validator.service.ts
@@ -42,7 +42,7 @@ export class ModelValidatorService {
     violationErrors.forEach((error: ViolationError) => {
       this.notificationsService.validationError({
         title: error.message,
-        message: error.fix[0], //TODO should be changed, when more fixes are available
+        message: error.fix.join('; '),
         link: error.focusNode,
         timeout: 5000,
       });
