+++ core/libs/editor/src/lib/editor-dialog/components/fields/entity/extends-field/extends-field.component.ts
@@ -178,7 +178,6 @@ export class EntityExtendsFieldComponent extends InputFieldComponent<DefaultEnti
       return;
     }
 
-    // const newAbstractEntity = new DefaultAbstractEntity(this.metaModelElement.metaModelVersion, urn, entityName, []);
     const newAbstractEntity = new DefaultEntity({
       isAbstract: true,
       name: entityName,
